Sync setting of node / item active
Items with a associated node uses Node#toggle! to ensure that the node
and items active states stay in sync. If there is no associated node the
item will just toggle it self.

Setting of node active/inactive has moved the right nav on
Admin::Nodes#edit to seperate the setting of node active/inactive from
the setting of node display in navigation lists.